+++ R/na.operations.R
@@ -6,6 +6,8 @@
 #'
 #' @param x An object
 #' @param label A character string denoting the label to set \code{NA}s to.
+#' @param first Logical; should the new label be the first level?
+#' @param ... Other arguments (not in use at this time).
 #'
 #' @seealso \code{\link{is.na}}, \code{\link{anyNA}}
 #' @author Ethan Heinzen
@@ -19,19 +21,24 @@ allNA <- function(x) all(is.na(x))
 
 #' @rdname NA.operations
 #' @export
-includeNA <- function(x, label)
+includeNA <- function(x, label, ...)
 {
   UseMethod("includeNA")
 }
 
 #' @rdname NA.operations
 #' @export
-includeNA.factor <- function(x, label = "(Missing)")
+includeNA.factor <- function(x, label = "(Missing)", first = FALSE, ...)
 {
   lvl <- levels(x)
   if(label %in% lvl)
   {
     warning('"', label, '" already appears in levels(x).')
+  } else if(first)
+  {
+    levels(x) <- c(label, lvl)
+    x[] <- levels(x)[as.integer(x) + 1L]
+
   } else levels(x) <- c(lvl, label) # don't use factor() here, in case you lose attributes
   x[is.na(x)] <- label
   x
@@ -39,7 +46,7 @@ includeNA.factor <- function(x, label = "(Missing)")
 
 #' @rdname NA.operations
 #' @export
-includeNA.character <- function(x, label = "(Missing)")
+includeNA.character <- function(x, label = "(Missing)", ...)
 {
   if(label %in% x) warning('"', label, '" already appears in x.')
   x[is.na(x)] <- label